Migrate core components - navbar, sidebar, and footer
