+++ src/core/footer/footer.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+class Footer extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <div className='footer'>
+      </div>
+    );
+  }
+}
+
+export default Footer;
