+++ src/core/footer/index.js
@@ -1,0 +1,15 @@
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router';
+
+const mapStateToProps = state => ({
+
+});
+
+const mapDispatchToProps = dispatch => ({
+
+});
+
+export default connect(
+  null,
+  null
+)(withRouter(Footer));
