+++ src/core/navbar/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import Navbar from './navbar';
-import { receiveQuery } from 'actions/query_actions';
-import { loginUser, receiveUser } from 'actions/oauth_actions';
+import { receiveQuery } from 'actions/query_actions'; // TODO: Update!
+import { loginUser, receiveUser } from 'common/modules/user/actions';
 
 const mapStateToProps = ({ user }) => ({
   user,
