+++ src/core/sidebar/contributor_section.jsx
