+++ src/core/sidebar/sidebar_item.jsx
