+++ src/modules/channel/components/channel.jsx
