+++ src/modules/channel/components/index.js
