+++ src/modules/channel/reducer.js
